Move DateHelpers to lib
CRM and WATER came out of licence-import, and both depend on functions in DateHelpers.
